Moved `MetaEventListenerFlags` to a dedicated file.
