+++ glare/src/engine/meta/meta_event_listener_flags.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+namespace engine
+{
+	struct MetaEventListenerFlags
+	{
+		bool events                : 1 = true;
+		bool component_creation    : 1 = true;
+		bool component_update      : 1 = true;
+		bool component_destruction : 1 = true;
+	};
+}
